refactor([issue-755]): split GoalDetailPanel into subcomponents and a useGoalDetail hook
